# lib/api-zod/src/generated/types/profileSetupInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Sports Match Prediction API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileSetupInput {
  /**
     * @minLength 2
     * @maxLength 50
     */
  displayName: string;
  username?: string;
}